Split oversized Swift files
