+++ Sources/Views/CaptureWindowView.swift
@@ -15,23 +15,23 @@ struct CaptureWindowView: View {
   let onCancel: () -> Void
   var onAddSubreddit: (CaptureFormDraft) -> Void = { _ in }
 
-  @Query(sort: \Project.name) private var projects: [Project]
-  @Query(sort: \Subreddit.sortOrder) private var subreddits: [Subreddit]
+  @Query(sort: \Project.name) var projects: [Project]
+  @Query(sort: \Subreddit.sortOrder) var subreddits: [Subreddit]
 
-  @State private var title: String = ""
-  @State private var text: String = ""
-  @State private var notes: String = ""
-  @State private var selectedProject: Project?
-  @State private var selectedSubreddits: Set<UUID> = []
-  @State private var links: [String] = []
-  @State private var newLinkText: String = ""
-  @State private var droppedFiles: [URL] = []
-  @State private var existingMediaRefs: [String] = []
-  @State private var removedMediaRefs: [String] = []
-  @State private var originalMediaRefs: [String] = []
-  @State private var showPreview: Bool = false
-  @State private var saveError: String?
-  @AppStorage(SettingsKey.defaultProjectId) private var defaultProjectId: String = ""
+  @State var title: String = ""
+  @State var text: String = ""
+  @State var notes: String = ""
+  @State var selectedProject: Project?
+  @State var selectedSubreddits: Set<UUID> = []
+  @State var links: [String] = []
+  @State var newLinkText: String = ""
+  @State var droppedFiles: [URL] = []
+  @State var existingMediaRefs: [String] = []
+  @State var removedMediaRefs: [String] = []
+  @State var originalMediaRefs: [String] = []
+  @State var showPreview: Bool = false
+  @State var saveError: String?
+  @AppStorage(SettingsKey.defaultProjectId) var defaultProjectId: String = ""
   var body: some View {
     VStack(spacing: 0) {
       titleBar
@@ -167,150 +167,4 @@ struct CaptureWindowView: View {
     .onAppear { populateFromMode() }
   }
 
-  private var titleBar: some View {
-    HStack {
-      Text(titleText)
-        .font(.system(size: 13, weight: .semibold))
-
-      Spacer()
-
-      Button("Cancel", action: onCancel)
-        .font(.system(size: 11))
-        .foregroundStyle(.secondary)
-        .buttonStyle(.plain)
-        .accessibilityLabel("Cancel capture")
-        .accessibilityIdentifier("captureWindow.cancel")
-
-      Button("Save", action: save)
-        .font(.system(size: 11, weight: .semibold))
-        .foregroundStyle(canSave ? AppColors.redditOrange : AppColors.redditOrange.opacity(0.4))
-        .buttonStyle(.plain)
-        .disabled(!canSave)
-        .padding(.leading, 6)
-        .accessibilityLabel("Save capture")
-        .accessibilityIdentifier("captureWindow.save")
-    }
-    .padding(.horizontal, 16)
-    .padding(.vertical, 12)
-  }
-
-  private var titleText: String {
-    switch mode {
-    case .create: "New Capture"
-    case .edit: "Edit Capture"
-    }
-  }
-
-  private func fieldSection<Content: View>(
-    _ label: String,
-    optional: Bool = false,
-    required: Bool = false,
-    @ViewBuilder content: () -> Content
-  ) -> some View {
-    VStack(alignment: .leading, spacing: 6) {
-      HStack(spacing: 4) {
-        Text(label)
-          .font(.system(size: 10, weight: .medium))
-          .foregroundStyle(.secondary)
-          .tracking(0.3)
-        if optional {
-          Text("(optional)")
-            .font(.system(size: 10))
-            .foregroundStyle(.tertiary)
-        }
-        if required {
-          Text("(required)")
-            .font(.system(size: 10))
-            .foregroundStyle(AppColors.redditOrange)
-        }
-      }
-      content()
-    }
-  }
-
-  private var canSave: Bool {
-    CaptureHelpers.canSave(
-      title: title,
-      text: text,
-      selectedSubredditCount: selectedSubreddits.count
-    )
-  }
-
-  private var saveRequirementsMessage: String? {
-    CaptureHelpers.saveRequirementsMessage(
-      title: title,
-      text: text,
-      selectedSubredditCount: selectedSubreddits.count
-    )
-  }
-
-  private var currentDraft: CaptureFormDraft {
-    CaptureFormDraft(
-      title: title,
-      text: text,
-      notes: notes,
-      selectedProjectId: selectedProject?.id,
-      selectedSubredditIds: selectedSubreddits,
-      links: links,
-      newLinkText: newLinkText,
-      mediaURLs: droppedFiles
-    )
-  }
-
-  private func save() {
-    guard canSave else { return }
-    let selectedSubs = subreddits.filter { selectedSubreddits.contains($0.id) }
-    saveError = nil
-    let didSave = onSave(
-      CaptureFormResult(
-        title: title.trimmingCharacters(in: .whitespacesAndNewlines).nilIfEmpty,
-        text: text.trimmingCharacters(in: .whitespacesAndNewlines),
-        notes: notes.trimmingCharacters(in: .whitespacesAndNewlines).nilIfEmpty, links: links,
-        project: selectedProject,
-        subreddits: selectedSubs,
-        mediaURLs: droppedFiles,
-        removedMediaRefs: removedMediaRefs
-      ))
-    if !didSave {
-      saveError = "Save failed. Check selected media files and try again."
-    }
-  }
-
-  private var editCaptureId: UUID? {
-    if case .edit(let capture) = mode { return capture.id }
-    return nil
-  }
-
-  private func populateFromMode() {
-    switch mode {
-    case .create:
-      if let initialDraft {
-        title = initialDraft.title
-        text = initialDraft.text
-        notes = initialDraft.notes
-        selectedProject = projects.first { $0.id == initialDraft.selectedProjectId }
-        selectedSubreddits = initialDraft.selectedSubredditIds
-        links = initialDraft.links
-        newLinkText = initialDraft.newLinkText
-        droppedFiles = initialDraft.mediaURLs
-      } else if let uuid = UUID(uuidString: defaultProjectId) {
-        selectedProject = projects.first { $0.id == uuid }
-      }
-    case .edit(let capture):
-      title = capture.title ?? ""
-      text = capture.text
-      notes = capture.notes ?? ""
-      selectedProject = capture.project
-      selectedSubreddits = Set(capture.subreddits.map(\.id))
-      links = capture.links
-      existingMediaRefs = capture.mediaRefs
-      originalMediaRefs = capture.mediaRefs
-    }
-  }
-}
-
-extension String {
-  fileprivate var nilIfEmpty: String? {
-    isEmpty ? nil : self
-  }
 }
